+++ app/system/internal/handler/member/team/info_handler.go
@@ -1,0 +1,31 @@
+// Code scaffolded by goctl. Safe to edit.
+// goctl 1.10.0
+
+package team
+
+import (
+	"net/http"
+
+	"github.com/zeromicro/go-zero/rest/httpx"
+	"ovra/app/system/internal/logic/member/team"
+	"ovra/app/system/internal/svc"
+	"ovra/app/system/internal/types"
+)
+
+func InfoHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		var req types.IdReq
+		if err := httpx.Parse(r, &req); err != nil {
+			httpx.ErrorCtx(r.Context(), w, err)
+			return
+		}
+
+		l := team.NewInfoLogic(r.Context(), svcCtx)
+		resp, err := l.Info(&req)
+		if err != nil {
+			httpx.ErrorCtx(r.Context(), w, err)
+		} else {
+			httpx.OkJsonCtx(r.Context(), w, resp)
+		}
+	}
+}